add resizeable cards
