+++ apps/web/src/components/(editor)/card.tsx
@@ -2,7 +2,7 @@ import { useHighlightContext } from "@/context/highlight";
 import { useCard } from "@/hooks/useCard";
 import { Scrycard } from "react-scrycards";
 
-export function Card({ id }: { id: string }) {
+export function Card({ id, width = 200 }: { id: string; width?: number }) {
     const card = useCard(id);
     const { pushSelected, setHovered, setOpen } = useHighlightContext();
 
@@ -17,7 +17,7 @@ export function Card({ id }: { id: string }) {
             }}
             className="overflow-clip"
         >
-            <Scrycard animated card={card} size="md" width="200px" />
+            <Scrycard animated card={card} size="md" width={`${width}px`} />
         </button>
     );
 }
